Item quantities changes to ints, phone numbers now 15 digits, address now have address2 (line 2), city, country

+++ cs3750Fall2013/RecreationOutlet/Warehouse/RecOutletWarehouse/RecOutletWarehouse/Models/INVENTORY.cs
@@ -9,8 +9,8 @@ namespace RecOutletWarehouse.Models
         public byte LocationID { get; set; }
         public long RecRPC { get; set; }
         public byte QtyTypeID { get; set; }
-        public short QtyOnHand { get; set; }
-        public short QtyThreshold { get; set; }
+        public int QtyOnHand { get; set; }
+        public int QtyThreshold { get; set; }
         public virtual ITEM ITEM { get; set; }
         public virtual LOCATION LOCATION { get; set; }
         public virtual QTY_TYPE QTY_TYPE { get; set; }
